+++ OISC_VM/IMemoryBus.cs
@@ -7,10 +7,10 @@ namespace OISC_VM
 {
     public interface IMemoryBus
     {
-        int ReadData(int memoryLocation);
-        void WriteData(int memoryLocation, int value);
-        int[] ReadDataRange(int rangeStart, int rangeLength);
+        long ReadData(long memoryLocation);
+        void WriteData(long memoryLocation, long value);
+        byte[] ReadDataRange(long rangeStart, long rangeLength);
 
-        InstructionOperands FetchInstrucitonOperands(int memoryLocation);
+        InstructionOperands FetchInstrucitonOperands(long memoryLocation);
     }
 }
